Clean up KafkaPublisherSpec
